[clang][NFC] Move CWG273 test into its own file
